+++ components/Main.vue
@@ -1,3 +1,0 @@
-<template>
-  <main>Body</main>
-</template>