+++ state/app.js
@@ -1,0 +1,7 @@
+import { reactive } from 'vue'
+
+export const app = reactive({
+  title: 'Fairy',
+  language: 'en',
+  description: 'A general IDE and computational notebook for everyone'
+})